feat(torznab/db): Reset cached indexer failures on startup (#616)
If an indexer failed with a non-429, it may be due to a configuration
issue. This is frustrating for new users because they are working on
getting their configuration correct and then if something breaks, they
have to wait 10 minutes until they can continue.

This PR clears out any saved UNKNOWN_ERRORs on startup, allowing users
to freely tinker with their configurations.