fix(client/rtorrent): send content-type text/xml
hotio reported that certain builds of rtorrent (particularly jesec's rflood) requires the text/xml for requests. if not provided, rtorrent will potentially return empty strings

rtorrent is the only client implementation in cross-seed that does not currently provide a content-type in the requests.

this adds the header "Content-Type: text/xml" to the call to clientCreator() to fix this issue.