bug(clients/qbit): Properly handle subfolder layout (#665)
Did some digging into how subfolder layout is handled. It should be
completely fixed now. Problem was if a torrent with a single file that
naturally had a folder, it would get the incorrect path.

Also getting torrent info in a way that doesn't rely on infohash
versions.

Potentially fixed bug with rechecking searchee instead of newTorrent in
deluge.

---------

Co-authored-by: zakary <zak@ary.dev>