fix(client/qbit): fix tagging erroneously for torrent-backed searchee
this fixes the case when path is truthy (linked) and the searchee is torrent backed

because we only check for path before setting category to eliminate autotmm problems by using only one category, the category for linking is set to linkCategory.

if the searchee is torrent-backed, we should be tagging from the category of the searchee, not the destination category of the candidate.