fix(regex): movie matched as episode (#687)
The movie `A.Whisker.Away.2020.1080p.BluRay.Opus5.1.H.265-LYS1TH3A.mkv`
was matched as an episode due to the `Opus5.1` where `s5` was season and
`1` was episode.

This matched due to the support for `SXX - E?YY` anime format. But it
turns out that `.` is never used instead of `-` or `_` here so it can
safely be removed.

Also found
`Haunted.Mansion.2023.1080P.Bluray.TrueHD.Atmos7.1.Hybrid.DoVi.HEVC.X265-FZHD`
where `Atmos7.1` has the same issue.