feat(arr-lookup): fetch category and id caps for torznab searching (#640)
closes #437


enhancement to search

- searches an added radarr or sonarr instances for respective movies or
tv in the local library/db (parse endpoint). parses out the appropriate
ids for searching
- improves the potential capabilities of media library searches
substantially in subsequent open PR, as well as foreign language
searches
    - requires indexer to have the id search caps listed
- only searches tv or movies ids on indexers with the corresponding
caps. no id searches on trackers that do not have it
- uses same "style" URL and queries as torznab, supports multiple
instances in string array
- reverts to old behavior if it cannot search ids or cannot find the
content in one of the arrs
- logging done in verbose/debug

---------

Co-authored-by: ShanaryS <shanarys@gmail.com>
Co-authored-by: Michael Goodnow <mmgoodnow@gmail.com>