enhancement(action/unlink): unlink failed injections and save incomplete torrents (#712)
The unlink logic was pulled from the `inject-saved-torrents` branch so
it has been used for a while without issues. I've expanded it to cover
the bug here [from
discord](https://discord.com/channels/880949701845872672/1228878509414420594/1250084947985694721).
Torrents with the same hash from different trackers would still be
linked with `flatLinking: false`.

I also starting saving torrents when source is incomplete, which is for
the inject feature but we can start doing this now.