fix(regex/ep): years cannot be episode num (#734)
This just prevents 19XX|20XX from being episode numbers. Which if course
prevents actual episodes of that number from matching, but that only
happens in case of anime absolute formatting. And very few anime are
that high. While it's much more common for the episode number to match a
year in two scenarios:
1. `EXX 2024`
2. SXX `2024`

https://regex101.com/r/i0igR5/6

This was not changed for the anime regex since it already parses years.