fix(qbit/inject): better handling of autoTMM and duplicate cats for injection (#736)
Using `savePath` and `downloadPath` is now correctly applied if
`!autoTMM`. Previously this was `path` which would be incorrect if the
user is not linking and does not have autoTMM.

autoTMM would not be enabled if the user had flatLinking false but was
not linking. If autoTMM is ever true that means we are not linking and
the original torrent has it, so we can just use the value directly.

Refactored how tagging is handled so the logic is much simpler. The
state of `searcheeInfo` is all we need.

We will now actually duplicate the category if we are not linking. This
preserves old v5 behavior so cross seeds don't show up in arr
categories. If it was linked, will will use tags as we have in v6.

The new logic was simplified by relying on what the state could be based
on `searcheeInfo` being valid and the values initialized by the `path`
ternary.