enhancement(reverselookup): using regex to parse rss/announce names instead of Fuse (#725)
This replaces the Fuse method of reverse lookup to leveraging the
existing regex. Only the title, season, episode is used. All other parts
should be handled in the decide stage. The title is cleansed as much as
possible then converted to lowercase and with symbols and spaces
removed. This is already much better than the Fuse, but we then add a
fuzzy using the levenshtein distance. The max distance is
`Math.floor(cleanTitle.length / 4)` so short titles needs to be exact
and longer titles get more leeway. I haven't come across a situation
where we need this fuzzy, but I'm sure they exist.

This method vastly outperforms Fuse in all aspects.
1. It's much faster. It's always ~50ms to get all matching items from
the database for me. Where Fuse took 150ms if nothing was found up to
multiple seconds if items were.
2. It has far less false positives that Fuse. Fuse would erroneously
move to decide on at least 1/20 announces for me, I haven't seen a
single false positive for the new method. They do exist, for example I'm
not using the year in the key for movies since it can be wrong, but of
course it's still far better than before.
3. Better true positives. Due to all the parsing, it handles weird
naming formats much better and doesn't get tripped up by spaces or dots
like Fuse.
4. A not insignificant benefit is that this is much more
"deterministic". You can easily make improvements and logically reason
about the algorithm. Where as the previous method was essentially a
black box with a few parameters to tweak.

Due to how much more robust this is, we now create searchees from all
matches rather than just the first. These are ordered by most files to
reduce the chance of partials. I'm also using the improved dupe check
originally from the `cache-searche-queries` branch.