enhancement(search): search unknown MediaType only if supported by indexer categories (#746)
This aims to prevent searching for media that can't exist on the indexer
for unknown media types as well.
* If the searchee contains a video, always search unknown media type
* Otherwise only search if the indexer has additional categories (ones
not explicitly recognized by cross-seed)

Also now matching `MediaType: MOVIE` for .rar releases. Before it was
just a logging thing and maybe some uncessary api calls to sonarr. But
is now necessary with this new behavior.