feat(decide/cache): retry cached decisions (#707)
Updated how decision cache is handled so that users will never need
delete the database or `cross-seed clear-cache` for v6 and beyond. For
most decisions, it will reassess as the cache could get invalidated in
many ways by config changes or future updates. To better assist this,
caching has been updated to take advantage of the
`inject-saved-torrents` feature of matching without snatching if the
torrent exists in `torrent_cache`. In addition, all .torrent files that
are snatched, not just successful matches, will now be saved in
`torrent_cache`. Cache .torrent files are retrieved by the candidate
guid only, so similar searchees don't need to resnatch. The priority of
decisions has also changed to better handle fast paths when cached. With
these changes, torrents are only ever snatched ONCE for the lifetime of
`cross-seed`.

With all the changes to decision, it would benefit from having true
unique entries so I opted to solve #522 with a migration. I added a
column for `fuzzy_size_factor` (`fuzzy_size_threshold` or
`seasonFromEpisodes`) as well. I also deleted all decisions that had a
null `info_hash` partly because it was necessary to not overload node's
memory but it also serves to give users a fresh start with all the
improvements with zero effect on `torrent_cache`.

The only decisions that won't be reassessed are
`RELEASE_GROUP_MISMATCH`, `PROPER_REPACK_MISMATCH`,
`RESOLUTION_MISMATCH`, `SOURCE_MISMATCH`, `MAGNET_LINK`. I would like
for us to try to finalize these regex's now as if we update it in the
future, we would need to clear them for users.

I removed checking `matchMode` in the decisions that won't be retried as
these decisions are permanent but matchMode might change. I added
checking the anime group to handle some edge cases where the group is
matched incorrectly. No longer checking REPACK/PROPER as we always want
to pass in risky/partial but we can't use matchMode for this either.

I also changed `filterTimestamps` so users won't need to change
`excludeRecentSearch` or `excludeOlder` when adding new indexers.
Otherwise `excludeOlder` keeps the behavior of checking for the first
time seen across any indexer.

---
With these changes, we can confidently state there is never a reason to
delete the database or cache for v6 and forever. `cross-seed
clear-cache` has been updated to only delete decisions where `info_hash`
is null. We no longer need to hurt the cache as these decisions will be
reassessed anyways. I removed clearing indexer caps here as it's updated
for all indexers at every startup.

If the user has updated their config, they would only need to possibly
update `excludeRecentSearch` and `excludeOlder`. If we modify the
algorithms for decisions that won't be retried, `cross-seed clear-cache`
will need to be ran. We can ask user to run it but I think we can do it
for them with our own migration logic. If we add the current release of
`cross-seed` to the settings table, we can check that at startup. We can
then perform simple housekeeping tasks between versions that aren't a
database migration.