enhancement(searchee): parse 'Season XX' or 'SXX' torrent titles (#754)
Fixes:
1. `SEASON_REGEX` no longer matches episodes
2. `SEASON_REGEX` matches titles that end with the season

Now parsing `Season XX` or `SXX` titles from torrent based searchees as
well. This required keeping track of the parsed title in the new `title`
field. `title` is equal to `name` in all other scenarios.

The two places where `name` is now being used is for linking and client
logging. Those require the original name. Otherwise `title` is better
for searching, logging, and `searchee.name` in database (different shows
won't collide).

As a bonus I added this for the anime releases that are just the show
name. So if the torrent title is `Show` and it has a single episode, we
will use the episode name as `Show SXXEYY`.

---------

Co-authored-by: zakary <zak@ary.dev>