update(mediatype/video): support for identifying mediatype of video disc images (#769)
this PR adds a new array for file extensions found on Blu-ray and DVD
disc images (excluding single image files - iso/img/etc) and adds a case
to the switch in getMediaType for unparsable titles/folders that contain
these to be identified as the VIDEO mediaType

in addition, all video discs will be rechecked and paused due to
inconsistencies possible. resolving these will require manual user
intervention if a conflict exists - the invalid files will need to be
deleted manually to break links - and then resumed. this will otherwise
overwrite the files and break the searchee's torrent.