feat(action/inject): add inject job to retry saved torrents and more (#670)
This feature allows cross-seed to try to inject torrents it finds in
outputDir. This means if something goes wrong and a match falls back to
saving, it will later be injected. The job is set to run hourly and is
instant when outputDir is empty. This can also be triggered by a new CLI
option `cross-seed inject`. This uses all the settings from the user's
config.js but if they want to inject torrents from another directory
they can use `cross-seed inject --inject-dir <dir_name>`.

This also works for torrents never seen by cross-seed, so a user could
add a torrent they got elsewhere and cross-seed will still inject it
provided it matched. They should follow the cross-seed format of
`[movie][Tracker]whatever.torrent`. This is because we use the tracker
when linking. If the format is not valid, will link to
`linkDir/UnknownTracker` if flatLinking is false.

This leverages existing functions so it shouldn't need to be updated if
improvements or features or added to cross-seed. It essentially goes
straight to the decide stage then action, never interacting with the
database. I also made it so it did no filtering. Since this won't be
requesting indexers, at the searchee creation stage and decide stage,
only the necessary checks are performed. `MATCH_PARTIAL` decisions are
ordered by torrent/data based over ensemble, then by most files to
maximize the partial percentage.

Since there is no filtering, it's possible for false positives to
happen. This won't apply to cross-seed retrying failed injections, as
those already went through the full pipeline, it's only possible if the
user supplied their own torrent to test. There is a warning on the help
for `cross-seed inject` and a warn message each time it's used. It's too
much detail to go into here, but I don't think any user will come across
a false positive because of this, the one I saw is an edge case of an
edge case. The no filtering does allow extra matches that cross-seed
otherwise would of rejected. Biggest being two season packs that has a
different number of episodes can be cross-seeded since we aren't
checking total size fuzzy. Essentially if you think something is a
cross-seed, `cross-seed inject` will definitively say so. We can use
this fact for support or future development.

## Behavior Changes

Will now update database when if a `TORRENT_NOT_COMPLETE` injection
result is found. Previously this wasn't done so that it will get retried
next search but this is no longer necessary.

Updated the risky and partial algorithms to uniquely use a searchee
file. So that multiple candidate files can't match with the same
searchee file. I knew about this when first implementing partial, but
it's only necessary now with the inject feature combined with the
`ensemble-matching` branch.

Now returning 200 for failed injections from announce since we will be
retrying those with the inject job. I'm also returning 200 if it already
exists. This allows autobrr to mark an announce as successful in case
rss or a webhook search injected between autobrr retries. Updated to
logging to state `Injected`, `Saved`, `Already Exists` instead of
`Added` along with status code.

Added a new decision `SAME_INFO_HASH`. In a few scenarios with the
inject feature, this needs to be handled separately from already exists.
It also helps prevent incorrect 200 response for announces if
radarr/sonarr added the torrent to the client before `cross-seed`
processed the announce.

Saved .torrent files also include info_hash so torrents with identical
names can be saved. This is necessary since we will be using `outputDir`
heavily.

We will no longer fall back to non-linking if it fails, we just saved
instead for later retrying. This only was possible in the case of a
match and the source wasn't found, pretty much always a user error.

I changed all `MATCH_PARTIAL` to save even on successful injection.
Also, `cross-seed inject` will not delete processed .torrent files
unless they are in client and complete or a week old. This was done
mainly in the case of `ensemble-matching` where a user could resume a
single torrent and remove the rest once that's complete. Then
`cross-seed inject` will automatically re-inject the others this time at
100% using the newly completed torrent. This not only allows users to
download the missing data once, but it gets linked between all so no
extra storage is used.

## Special Abilities

The inject feature also can link files from multiple searchees and
trigger a recheck in client to automate the `MATCH_PARTIAL` case above.
So a user would only need to resume a single partial match and the next
time inject is ran it will dynamically link the new files for the other
torrents (already exists in client, no re-injection) and performs a
recheck. No other user user input required aside from choosing which
tracker they want to incur the partial download on.

I also added the ability to inject torrents if the source is stalled.
These will always be treated as partial, so paused and rechecked. If all
the sources are `TORRENT_NOT_COMPLETE` and it hasn't been injected for a
day, it will be added with whatever files are present if linking is
enabled. This allows the user to choose to resume this "cross seed" so
they can potentially revive the stalled torrent. Of course this is not
really cross seeding, but it has the potential to revive dead torrents
so I think it's worth having.

---------

Co-authored-by: Michael Goodnow <mmgoodnow@gmail.com>