fix(config/seasonFromEpisodes): better validation (#851)
In case of `seasonFromEpisodes: undefined` (or missing from config), it
will now default to `null`, preventing validation errors. Even though a
default of `1` is fine for most users, the error has confused others.

Will now warn if using `action: save` due to the complexity of these
matches.

Ensemble requires clients to be connectable if torrentDir is enabled,
this will now be checked during validation.