update(blocklist/deluge): change labels to be category rather than tag (#849)
Labels were previously associated with the `tag:` prefix in blocklist,
the rationale was that there might be a time when labels are able to be
applied in more than one and the type was a string.

This will not be the case, Deluge uses labels in the same way that qBit
does, for save path, bandwidth and queue, etc. Multiple labels will not
ever be a concern.

This changes labels to category. Tags are assigned empty array, since no
tags exist in Deluge.


- [x] determine if No Label (and Uncategorized in qBit) need a constant
to distinguish them, both return `""` for this assignment (or rather
lack of)

---------

Co-authored-by: ShanaryS <shanarys@gmail.com>