feat(clients): option to use APIs instead of torrentDir (#864)
Fixes: #365 

Added `useClientTorrents` which replaces `torrentDir`. They cannot both
be enabled.

There is now a new type of Searchee, `SearcheeClient`. This is like
`SearcheeWithInfoHash`, but has a new field `savePath` which makes
linking much easier.

This change means that we now support renames in client or any
modification a user can do. Right before linking, we refresh the
searchee to catch any renames as the db only updates on added/deleted.
All searchees are also refreshed at the start of each search cadence.

Indexing torrents and data now happens asynchronously. There is also now
prevention of multiple concurrent indexes to not overload the
client/disk. If an index is in progress, announce/rss/webhoook will
simply wait on its completion without triggering a new one.

I think we should store the client and data indexes in the physical db
to prevent any chance of memory issues. However I think that should be
saved for the PR supporting multiple clients, as that implementation
will affect the schema. Considering that this feature is opt-in for
existing users, memDB should be fine until then. On a similar note, the
indexing functions can be refactored but should be saved for then.

---------

Co-authored-by: Michael Goodnow <mmgoodnow@gmail.com>