enhancement(preFIlter): exclude episode/season from arr directory detection (#900)
These can never be the arr dirs that we are trying to filter against.
This will only happen if the user has non-standard naming for their
arrs.

The other time this can happen is if it's trying to search `Specials`,
so now I added an explicit check for this. `Specials` will usually get
parsed into `Show S0` due to the episode files in the folder.