update(matchMode): use "strict" and "flexible" instead of "safe" and "risky" (#911)
This updates the names for `matchMode` to be less misleading. The new
ones are `strict` `flexible` `partial`. This better conveys that these
options are equally safe as many users took `safe` and `risky` too
literally.

I also think we should update the config template to `matchMode:
"partial"` and `seasonFromEpisodes: 0.8`. Users have long asked for
`optimal` settings and this should be close to it for most. If we are
assuming that users will be setting up linking, similar to how the
current options assume a client will be configured, we should start with
these options. If a user cannot use linking or they do not want to, the
options become opt-out for new users.

I think we have enough evidence that these options bring the user more
matches with no real increases in risk. Though I understand if we still
want to be conservative with this.