fix(cmd): only add options to relevant commands (#907)
Options such as `searchCadence` should only defined in daemon mode, so
that `cross-seed search --no-exclude...` doesn't require setting
`searchCadence: null` in the config before hand.

I also updated the `cross-seed diff` command to only show the files.
Since this command is to see why two torrents are not a match, `name`
and `length` are distracting. If the user wants to see the these, they
can use `cross-seed tree` separately (which also displays the parsed
title). The files are also sorted in different ways to make the diff
easier to read.