feat(api): add endpoint to trigger jobs (#909)
This is from speaking with @zakkarry about `cross-seed search
--no-exclude...`. Since we can't run two instances of `cross-seed` at
the same time (commands that affect the db), triggering the jobs in
daemon mode makes sense.

This follows up on the webhook params on overriding certain config
options for this request, only `excludeOlder` and `excludeRecentSearch`.

When a job is manually triggered, it will run now and it will be
scheduled at `cadence * 2` in the future. So the next search would be in
`2 days` instead of `1 day`. It will also prevent triggering the job
until we are within the cadence of the job. So the user would only be
able to trigger again once the scheduled run is `1 day` away.

Checks are done to see if the job is available to be ran (e.g. no
`searchCadence`) and if it's already running. If the job trigger was
successful but cross-seed was stopped before completion, it will been as
a normal scheduled run on startup. No history of the trigger happening
will exist as if the command was never sent.