update(outputDir): default to /config/cross-seeds (#916)
Users frequently misconfigure `outputDir` in a few ways:
- It's to where the client copies .torrent files
- It's used as a watch directory
- Programs such as qbit_manage delete the folder
- rTorrent does not have access to it with the same mount

These are solved by `outputDir: null`. This will be set to
`join(appDir(), "cross-seeds")` so it works for npm users and all
platforms.

To solve the rTorrent issue, it needed to be decoupled from outputDir.
`load.normal` and `load.start` have been replaced with `load.raw` and
`load.raw_start`. These are the same methods for injecting, but take a
base64 encoded .torrent instead of the path. I will ping the rTorrent
role in pre-release for testing, but there should be no issues. This
method as been around since 2007 and the arrs use it as well.

This means that we can get rid of the `:/cross-seeds` mount. Now users
only need a single mount to use cross-seed, `:/config`, along with other
mounts to their data if they are using linking. This will make setup for
first time users easier and eliminate all support issues around
`outputDir`.