enhancement(qbit): support `Don't create subfolder` (#922)
This adds support for the `Don't create subfolder` option in qbit which
means `cross-seed` can now handle all modifications that a user can do.
This is only supported with `useClientTorrents`, doing this for
`torrentDir` would require a large refactor of how we link which is why
it hasn't be supported all this time. This option removes the root
folder, so episodes are stored directly into the `savePath`, instead of
`savePath/Show.S01`.

This builds on #908 by now completely getting rid of `sourceRoot`, it
was only used to test if the file exists which is still being performed.
We now use `savePath/files[x].path` everywhere to get the absolute paths
of torrents.

I also added two more validation checks that moves implicit runtime
errors to startup:
- Require v4.3.1+ (released 11/25/2020) for qbit, this has been
implicitly required since v6 pre-release. It would fail at startup or
when trying to inject with an non-obvious error. Any users below this
version cannot use `cross-seed` currently, this let's them know why.
- Fail validation if a user is using `torrentDir` with `Don't create
subfolder` torrents. We have never supported this, and these users are
running into errors with these after injection. We can just log an error
for this though, it can now detect these and log it before inejction.