framework for multiple client support (#904)
Only a single client is supported on this branch.

The goal for this feature is to source searchees from multiple clients
and cross seed them to their respective clients. cross seeding from one
client to another is not supported. Injections will only happen in a
single client, the already exists checks looks at all clients.

The client config options have all been deprecated and `torrentClients:
[]` is used instead. Each url must be prefixed by it's client type
similar to blockList. This allows the user to define a `clientPriority`
which affects which client is chosen when source torrents are
duplicated/ambiguous.

The `URL(url).host` is used to distinguish between clients, duplicates
are not allowed. This is added to `SearcheeClient`. Logging for the
clients now includes this to distinguish themselves.