fix(linkDirs): fallback to temp file to test linking (#927)
We only try to test linking with `ALL_EXTENSIONS` to avoid using OS
files which will fail. Anything in `ALL_EXTENSIONS` should be fair to
use. If none were found, the test for that path was skipped and no
errors would occur.

Now we will create a file to test linking with if no existing ones are
found. This will now throw an error if the user has a misconfigured
empty drive.

The main reason for this is if the user has a drive with only games, we
may not find a file to test with. But `getLinkDir` will fail when
actually trying to link something unlike `testLinking`.