feat(clients): support multiple clients with dataDirs (#937)
Required a large refactoring of `action.ts` and which affected
`inject.ts` and the clients. Mainly was just about moving lots of the
work into their own separate functions to get more flexibility. Also
took more advantage of union types to help specify what's possible in
certain conditions.

All the new functionality is in `getClientAndDestinationDir()` with the
integration in `performActionsWithoutMutex()`. The changes in
`inject.ts` are simply due to the new function signatures in
`action.ts`. The changes for the clients are to standardize the handling
of `TORRENT_NOT_COMPLETE` and using `destinationDir` instead of `path`
for the `client.inject()` argument.

As before, a single client has the same behavior. `dataDirs` and
ensemble `dataDirs` choose a client by testing linking to the
`savePaths` for each client. It chooses the first client with a
successful link then finds a `linkDir`. Non-linking is not supported for
dataDirs. In principle there is no reason why it can't, but it would
require interacting with the filesystem which we should avoid for
non-linking setups. Interacting with the filesystem won't be necessary
with #934.

Users will likely desire some control over what goes where, but that's
for #934 to address.