fix(regex/group): prevent group matches from titles (#950)
If the title of the media contains a hyphen, there is a chance that it
matches as the release group. This would only be in cases of parsed
titles for ensemble `Re:ZERO -Starting Life in Another World.S3`.

This checks if the release group parsed from the title is a substring of
the one parsed from the entire release.