fix(torrentClients): treat readonly searchees as data-based (#956)
If `infoHash` is present in the `inject()` function, the current logic
expects it to be in client. By treating it as data-based, we only lose
`duplicateCategories` functionality. But this is not a big deal since
different clients uses `categories`, `tags`, and `labels` differently.
Rewriting the `inject()` functions to preserve this is not worth it.