enhancement(job/inject): trigger inject job on webhook completion (#957)
This improves the inject job workflow as we often want to run after a
download is completed in client:
1. The torrent that just completed had saved cross seeds due to
`TORRENT_NOT_COMPLETE`. The inject job now triggers within the next
minute to inject these earlier.
2. The torrent that just finished was a partial match. This could have
downloaded new files for other partials matches, mainly for this [FAQ
entry](https://www.cross-seed.org/docs/basics/faq-troubleshooting#my-partial-matches-from-related-searches-are-missing-the-same-data-how-can-i-only-download-it-once)

The next run time will be at its normal cadence unlike with the `/job`
api.