fix(testLinking): add more checks around creating temp file (#958)
Creating a temp file for testing linking is used if we can't find any
valid file to link with (the extensions we support) and `stat.dev` isn't
available. If we fail to create a tempfile, we log and error but do not
fail validation.
1. Check if we have write access to create a temp file. This can happen
if we are testing `dataDirs` as we only require read permissions for
that.
2. Verify tempfile was created before trying to use it, the write can
silently fail it seems.