update(db): remove memDB (#961)
fixes #910

1. New migration for `client_searchees`, `data`, and `ensemble` tables.
2. Update the `cleanup` job to also handle these tables to ensure
`client_searchees` are refreshed (catch renames, new categories, tags,
trackers, etc), and that `data` and `ensemble` gets pruned for paths
that no longer exists (only happens if `cross-seed` was down so files
weren't being watched).
3. Upon startup, we will clear tables for the options the user has not
configured. e.g. `torrent` gets cleared if no `torrentDir`, `data` is
clear if no `dataDirs.length`, etc.
4. Added `cross-seed clear-client-cache` to clear the `torrent`,
`client_searchees`, `data`, and `ensemble` tables. Only needed if the
user doesn't want to want on the cleanup job for the refresh or if they
moved to a new client with the same host.
5. No longer refreshing searchees at the start of every search, this
will now be handled by the `cleanup` job (at the start of the action
stage, we refresh the searchee so there is no chance of outdated
names/save_path causing a failed injection).