feat(inject/ignoreFiles): ignore certain non-relevant files for resume behavior with partial matches (#940)
The behavior idea I had in mind involves excluding known non-relevant
files as a second measure to check if resuming is desired. This should
cover most instances where any sample or miscellaneous file commonly
included in some trackers prevents resuming and provides a larger buffer
to resume with.

The option to enable this behavior is `ignoreNonRelevantFilesToResume`.
Defaults, if not included in the config (onboarding), to false. The
previous behavior stays intact with the 50MB limit, but I'm not sure
that increasing it as @ShanaryS suggests in #939 is necessary given this
option's functionality, and I would contend that that is not the best or
proper "total" solution for the job.

---------

Co-authored-by: ShanaryS <shanarys@gmail.com>