enhancement(cmd/inject): add `--ignore-titles` to skip fuzzy matching (#966)
Occasionally there might be torrents that are matches but fail the
`cross-seed inject` fuzzy lookup. This adds a command to bypass this
check if the user is confident the torrents are simply named badly.

This defaults to false (current behavior) and is only available with the
`cross-seed inject` command. If a potential match was rejected due to a
title mismatch, the torrent will not be deleted.