refactor(async): replace fs sync functions with async (#967)
This ends up being a pretty large refactor. The goal is to replace all
the `fs.*Sync` with their async versions as mentioned a while back by
@mmgoodnow. I organized `utils.ts` and when through a lot of the async
functions with `eslint` rules to clean up their usages.

1. Reorganized utils
2. Use `fs/promises` instead of `*Sync` function calls in `fs`. The
exceptions are for zod schema parsing which of course is low value. It
would require `parseAsync` which seems to have performance issues
reported at the very least.
3. Created and used `mapAsync`, `flatMapAsync`, `reduceAsync`,
`findAsync`, `everyAsync`, and `someAsync` functions in addition to the
existing `filterAsync` to standardize async usage for functional
pipelines.
4. Use `Uint8Array` instead of `Buffer` for saving .torrent files.
`Buffer` has been essentially been deprecated for JS. `Metafile` still
returns `Buffer` as we use some of the functions for injecting into
clients.
5. For `injectDir`, check against `undefined` to know if the user is
running the command `cross-seed inject` for the warning logs.
6. Reduce the stress that the `rss job` has on the event loop by adding
a `100ms` wait between candidates. Lots of API requests would fail or
other operations freeze during the duration of the `rss job`. It's still
not perfect as even `/api/ping` has a noticeable delay while the
candidates are being processed or announces flooding in after the job is
complete.
7. Update `withMutex` usage per
https://github.com/cross-seed/cross-seed/pull/868#discussion_r1920717798
to reduce the chance of user error. Instead of the caller handling the
mutex, the functions that need it will always use it. It's not perfect
but we need the flexibility of the footgun.
8. Add extra checks to `eslint` to require more explicit usage of async
functions and variables. This should help preventing a significant
amount of bugs that can be introduced from `sync` to `async` transition
or vice versa. Things like non-awaited promises will need `void
asyncFunc()` to show it's itentional, await must be used in places like
`try/catch` so errors are handled, await cannot be used on sync
operations, no unnecessary type assertions (e.g `neverNullishVar!`),
etc...